move setup() and structopt to setup.rs
