+++ src/main.rs
@@ -1,22 +1,14 @@
-extern crate atty;
-extern crate humantime;
-extern crate regex;
-extern crate structopt;
-extern crate subprocess;
-extern crate tempfile;
+mod setup;
 
 use std::env;
-use std::f64;
 use std::fs::File;
 use std::io::prelude::*;
-use std::io::{self, BufRead, SeekFrom};
+use std::io::SeekFrom;
 use std::process;
 use std::thread;
-use std::time::{Duration, Instant, SystemTime};
+use std::time::{Instant, SystemTime};
 
-use humantime::{parse_duration, parse_rfc3339_weak};
-use regex::Regex;
-use structopt::StructOpt;
+use setup::{setup, ErrorCode, Opt};
 use subprocess::{Exec, ExitStatus, Redirection};
 
 static UNKONWN_EXIT_CODE: u32 = 99;
@@ -55,32 +47,6 @@ fn main() {
         state.summary,
         state.tmpfile,
     )
-}
-
-fn setup() -> (Opt, Vec<String>, usize, Instant) {
-    // Time
-    let program_start = Instant::now();
-
-    // Load the CLI arguments
-    let opt = Opt::from_args();
-    let count_precision = Opt::clap()
-        .get_matches()
-        .value_of("count_by")
-        .map(precision_of)
-        .unwrap_or(0);
-
-    // Number of iterations
-    let mut items: Vec<String> = opt.ffor.clone().unwrap_or_else(|| vec![]);
-
-    // Get any lines from stdin
-    if opt.stdin || atty::isnt(atty::Stream::Stdin) {
-        io::stdin()
-            .lock()
-            .lines()
-            .for_each(|line| items.push(line.unwrap().to_owned()));
-    }
-
-    (opt, items, count_precision, program_start)
 }
 
 struct State {
@@ -257,7 +223,7 @@ fn counters_from_opt(opt: &Opt, items: &[String]) -> Vec<f64> {
     } else if !items.is_empty() {
         items.len() as f64
     } else {
-        f64::INFINITY
+        std::f64::INFINITY
     };
     loop {
         start += step_by;
@@ -319,140 +285,6 @@ impl StringFromTempfileStart for String {
         tmpfile.seek(SeekFrom::Start(0)).ok();
         tmpfile.read_to_string(&mut stdout).ok();
         stdout
-    }
-}
-
-#[derive(StructOpt, Debug)]
-#[structopt(
-    name = "loop",
-    author = "Rich Jones <miserlou@gmail.com>",
-    about = "UNIX's missing `loop` command"
-)]
-struct Opt {
-    /// Number of iterations to execute
-    #[structopt(short = "n", long = "num")]
-    num: Option<f64>,
-
-    /// Amount to increment the counter by
-    #[structopt(short = "b", long = "count-by", default_value = "1")]
-    count_by: f64,
-
-    /// Amount to offset the initial counter by
-    #[structopt(short = "o", long = "offset", default_value = "0")]
-    offset: f64,
-
-    /// How often to iterate. ex., 5s, 1h1m1s1ms1us
-    #[structopt(
-        short = "e",
-        long = "every",
-        default_value = "1us",
-        parse(try_from_str = "parse_duration")
-    )]
-    every: Duration,
-
-    /// A comma-separated list of values, placed into 4ITEM. ex., red,green,blue
-    #[structopt(long = "for", parse(from_str = "get_values"))]
-    ffor: Option<Vec<String>>,
-
-    /// Keep going until the duration has elapsed (example 1m30s)
-    #[structopt(
-        short = "d",
-        long = "for-duration",
-        parse(try_from_str = "parse_duration")
-    )]
-    for_duration: Option<Duration>,
-
-    /// Keep going until the output contains this string
-    #[structopt(short = "c", long = "until-contains")]
-    until_contains: Option<String>,
-
-    /// Keep going until the output changes
-    #[structopt(short = "C", long = "until-changes")]
-    until_changes: bool,
-
-    /// Keep going until the output changes
-    #[structopt(short = "S", long = "until-same")]
-    until_same: bool,
-
-    /// Keep going until the output matches this regular expression
-    #[structopt(short = "m", long = "until-match", parse(try_from_str = "Regex::new"))]
-    until_match: Option<Regex>,
-
-    /// Keep going until a future time, ex. "2018-04-20 04:20:00" (Times in UTC.)
-    #[structopt(
-        short = "t",
-        long = "until-time",
-        parse(try_from_str = "parse_rfc3339_weak")
-    )]
-    until_time: Option<SystemTime>,
-
-    /// Keep going until the command exit status is non-zero, or the value given
-    #[structopt(short = "r", long = "until-error", parse(from_str = "get_error_code"))]
-    until_error: Option<ErrorCode>,
-
-    /// Keep going until the command exit status is zero
-    #[structopt(short = "s", long = "until-success")]
-    until_success: bool,
-
-    /// Keep going until the command exit status is non-zero
-    #[structopt(short = "f", long = "until-fail")]
-    until_fail: bool,
-
-    /// Only print the output of the last execution of the command
-    #[structopt(short = "l", long = "only-last")]
-    only_last: bool,
-
-    /// Read from standard input
-    #[structopt(short = "i", long = "stdin")]
-    stdin: bool,
-
-    /// Exit with timeout error code on duration
-    #[structopt(short = "D", long = "error-duration")]
-    error_duration: bool,
-
-    /// Provide a summary
-    #[structopt(long = "summary")]
-    summary: bool,
-
-    /// The command to be looped
-    #[structopt(raw(multiple = "true"))]
-    input: Vec<String>,
-}
-
-fn precision_of(s: &str) -> usize {
-    let after_point = match s.find('.') {
-        // '.' is ASCII so has len 1
-        Some(point) => point + 1,
-        None => return 0,
-    };
-    let exp = match s.find(&['e', 'E'][..]) {
-        Some(exp) => exp,
-        None => s.len(),
-    };
-    exp - after_point
-}
-
-#[derive(Debug)]
-enum ErrorCode {
-    Any,
-    Code(u32),
-}
-
-fn get_error_code(input: &&str) -> ErrorCode {
-    if let Ok(code) = input.parse::<u32>() {
-        ErrorCode::Code(code)
-    } else {
-        ErrorCode::Any
-    }
-}
-
-fn get_values(input: &&str) -> Vec<String> {
-    if input.contains('\n') {
-        input.split('\n').map(String::from).collect()
-    } else if input.contains(',') {
-        input.split(',').map(String::from).collect()
-    } else {
-        input.split(' ').map(String::from).collect()
     }
 }
